change signature for disasemble

+++ disassemble.cc
@@ -34,7 +34,7 @@ const std::string &getGPRName(uint32_t r) {
 
 
 
-void disassemble(std::ostream &out, uint32_t inst, uint32_t addr) {
+void disassemble(std::ostream &out, uint32_t inst, uint64_t addr) {
   out << getAsmString(inst,addr);
 }
 
@@ -70,7 +70,7 @@ void stopCapstone() {
   cs_close(&handle);
 }
 
-std::string getAsmString(uint32_t inst, uint32_t addr) {
+std::string getAsmString(uint32_t inst, uint64_t addr) {
   std::stringstream ss;
   cs_insn *insn = nullptr;
   size_t count = cs_disasm(handle,reinterpret_cast<const uint8_t *>(&inst),
